Finished the neural network.  I commented out the 2-player code to run my neural network.

+++ MP4_Reinforcement_Learning/loader.py
@@ -42,7 +42,7 @@ def loadFile(fileName):
             break
 
     states = data[:,:-1]
-    actions = data[:,-1]
+    actions = (data[:,-1]).astype(np.int32)
     return (states, actions)
